==> src/00_demographic_data_wrangling.Rmd <==
---
title: "Metropolitan Statistical Area Evaluation"
author: "Grace Bianchi"
date: "5/12/2022"
output:   
  html_document: 
  code_folding: hide
---

This report provides an explanatory overview of demographic and real estate data for metropolitan statistical areas. The code wrangles and analyzes data collected from government agencies to create metrics to explore the relationships between criteria and compare across MSAs. Specifically, this data analysis includes population growth, household income, employment rates, house prices, and property taxes. Preliminary analysis of population change and size narrowed down potential areas which experienced population growth rate greater than 0.5% and have populations greater than 300,000 people. In addition, 27 MSAs were removed from analysis, as requested by the client, due to hyper-competitive real estate markets. In brief, there are 29 metropolitan statistical areas with potential for ZNE Capital aqcuire real estate.

## Preliminary analysis {.tabset}

```{r setup, include=TRUE, warning = FALSE, message = FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)

library(tidyverse)
library(here)
library(janitor) #clean_names
library(DT) # create data table
# library(scales) 
library(lubridate)
library(gridExtra) 
library(ggrepel)
library(kableExtra)
library(googlesheets4) # output tables to google sheet
library(tigris) # imports Census shape files
library(sf)
options(tigris_class = "sf")
options(tigris_use_cache = TRUE)
library(openxlsx)
```

### Population Growth

All states were considered with the exception of CA, HI, and AK. Initial analysis of Metropolitan Statistical Areas (MSAs) relied on Census data of annual population estimates, in which the population had to be greater than 500,000 individuals for the purpose of attracting investors. Population change from 2020 to 2021 was calculated by $$\frac{2021 estimate- 2020 estimate}{2020 estimate} * 100$$ MSAs with a population growth \> 0.5%, were selected to gather further data on. The 0.5% growth was determined.

Dataset: Annual Resident Population Estimates and Estimated Components of Resident Population Change for Metropolitan and Micropolitan Statistical Areas and Their Geographic Components: April 1, 2020 to July 1, 2021 (CBSA-EST2021-ALLDATA)

Source: [U.S. Census Bureau, Population Division](https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-metro-and-micro-statistical-areas.html)

```{r population growth data wrangling}
# read in the data MSA population data from Census
msa_pop_2021_data <- read_csv(here("data/raw/population_est2021.csv")) %>% 
  clean_names() %>% # data wrangling to make column headers in lower_snake_case
  filter(lsad == "Metropolitan Statistical Area") %>% # select MSA cities
  select("cbsa", "name", "popestimate2020", "popestimate2021", "residual2020", "residual2021") %>% # select data 
  separate(col = "name",
           into = c("MSA", "state"), # split city and state data
           sep = ", ") %>% 
  mutate(pop_percent_change = (popestimate2021-popestimate2020)/popestimate2020 * 100)  # calculate population % change 

# create object with viable MSAs for client 
potential_MSAs <- msa_pop_2021_data %>% 
  filter(pop_percent_change > 0.5 ) %>% # select areas with population growth greater than .5%
  filter(popestimate2021 >= 500000) %>% 
  mutate(pop_percent_change20_21 = round(pop_percent_change, digits = 3)) %>% 
  filter (!state %in% c("CA", "HI", "AK")) %>% # remove areas within these states
  filter(!cbsa %in% c("10580", "10900", "12420", "12940", "14260", "15980", "18580", "19100", "19660", "19740", "23540", "25420", "26420", "27260", "29820", "34940", "36260", "36740", "38060", "39100", "39300", "39340", "41620", "45220", "45300", "46060", "46140", "39900")) %>%  # remove MSAs as instructed by client 
  select(cbsa, MSA, state, pop_percent_change20_21, popestimate2021) 

unfiltered_msas <- msa_pop_2021_data %>% 
  filter(pop_percent_change>0.5) %>% 
  select(1:3, pop_percent_change)
```

```{r population growth map, eval=FALSE}
## optional visualization
# read in MSA shapefiles from Census
cbsa <- core_based_statistical_areas(cb = TRUE) %>% 
  mutate(CBSAFP = as.numeric(CBSAFP)) %>% 
  inner_join(potential_MSAs, by = c("CBSAFP" = "cbsa")) # merge msa data with shape files

# state and U.S. map coordinates
states <- map_data("state")
usa<- map_data("usa") 

# map of U.S. with MSAs highlighted
ggplot() +
  geom_polygon(data = states,
               aes(x = long, y = lat,
                   group = group)) + 
  geom_sf (data = cbsa, fill = "lightgreen") +
  labs(title = "Metro Areas included in this analysis", x = "", y = "")
```

Figure 1. Population growth in Metropolitan Statistical Areas (n=29) in 2020-2021. Highlighted areas have a population \> 300,000 individuals and a population growth rate greater than 0.5%. The colors of the MSAs represent incremental growth rates, in which low(green) is 0.5-1% increase, medium(blue) is 1-2% increase, and high(pink) is 2-3.5% increase. Data: U.S. Census Bureau, Population Division.

### Household Income

Household Median Income data was also obtained from Census data at the MSA level and amended to existing MSAs with a positive population increase. Median is a better indicator of the population than the mean income values due to the positive skew from the highest incomes.

Dataset: Median Household Income (USD) INCOME IN THE PAST 12 MONTHS (IN 2021 INFLATION-ADJUSTED DOLLARS)

Source: [U.S. Census Bureau](https://data.census.gov/cedsci/table?q=median%20household%20income&t=Income%20and%20Earnings&g=0100000US%243100000&tid=ACSST5Y2020.S1901)

```{r household income data wrangling}
MSA_income_data <- read_csv(here("data/raw/income_data2021.csv")) %>% 
  row_to_names(1) %>% 
  clean_names() %>% 
  mutate(median_income = as.numeric(estimate_households_median_income_dollars),
  income_moe_USD = as.numeric(margin_of_error_households_median_income_dollars)) %>% 
  separate(col = "geography",
           into = c("id", 'cbsa'),
           sep = 9) %>% 
  mutate(cbsa = as.numeric(cbsa)) %>% # change column class from chr to dbl
  select(cbsa, median_income)

  
# join household median income to existing msa dataframe
potential_MSAs <- potential_MSAs %>% 
  inner_join(MSA_income_data)


unfiltered_msas <- unfiltered_msas %>% inner_join(MSA_income_data)
```

```{r, household income visual, fig.height = 7, fig.align='center', eval = FALSE}
ggplot(MSA_house_income, aes(x = median_income, y = reorder(MSA, median_income))) +
  geom_errorbarh(aes(xmin = median_income - income_moe_USD, xmax = median_income + income_moe_USD)) + 
  geom_point(size = 3, color = "darkgreen") + 
  theme_minimal(base_size = 12.5) + 
  labs(title = "Median household income in 2021", 
       subtitle = "", 
       x = "estimate (in USD)", 
       y = "") + 
  scale_x_continuous(labels = label_dollar())
  
```

Figure 2. Median Household income of Metropolitan Statistical Areas in 2021. The green points indicate the median income value for each MSA with the margin of error presented by the bars on either side of each point. Data: U.S. Census Bureau.

### Employment Rates

MSA monthly unemployment rates (seasonally-adjusted) were provided by The Bureau of Labor statistics in the civil labor force data at the MSA level. The average annual unemployment rates were calculated using the monthly rates from 2012 to 2020. Employment rates were calculated by dividing the number of employment / civilian labor force. The employment data for the month of January in years 2018-2021 was selected to calculate the change in employment rates over time. Year-over-Year employment growth was calculated by dividing the employment rate of the year by the previous year and subtracting one; for example, 2018 employment growth = (2019 employment rate / 2018 employment rate) - 1.

Annual Average Employment Rates & Year-over-Year change civil labor force data from [The Bureau of Labor Statistics](https://www.bls.gov/lau/metrossa.htm)

Monthly unemployment rates (seasonally-adjusted) from [The Bureau of Labor Statistics](https://www.bls.gov/web/metro/metro_area_3mma.pdf)

```{r, employment data wrangling}
unemployment_data <- read_csv(here("data/raw/unemployment_data.csv")) %>% #rename col headers
  clean_names() %>% 
  rename(cbsa = x3,
         area = x4,
         year = x5,
         month = x6,
         civilian_labor_force = x7,
         employment = x8,
         unemployment = x9,
         unemployment_rate = x10) %>% 
  drop_na() %>% 
  mutate(employment_rate = (employment/civilian_labor_force)*100) %>%  # calculate
  select(cbsa, area, year, month, employment_rate)

# data wrangling, only selecting data for potential MSAs
MSA_employment <- unemployment_data %>% 
  filter(year >= 2018) %>% 
  mutate(cbsa = ifelse(cbsa == '76750', "38860", cbsa)) %>% # change south portland, ME to corresponding cbsa
  mutate(cbsa = as.numeric(cbsa)) %>% 
  group_by(area, cbsa, year) %>% 
 #  right_join(potential_MSAs) %>% 
  summarize(avg_employment_rate = mean(employment_rate)) %>% 
  mutate(avg_emp_rate = (avg_employment_rate-lag(avg_employment_rate))/lag(avg_employment_rate)*100, 
          avg_emp_change = round(avg_emp_rate, digits = 4)) %>% 
  filter(year=="2021") %>% 
  select(cbsa, area, avg_emp_change)
  
  
              
client_MSAs <-potential_MSAs %>% 
  left_join(MSA_employment)

unfiltered_msas <- unfiltered_msas %>% 
  left_join(MSA_employment)
  
```

```{r employment visuals, fig.align = "center", fig.show="hold", fig.height= 7, fig.width=10, eval = FALSE}
# visual of MSA employment rates, shows major drop during covid
ggplot(MSA_employment, aes(x = year, y = avg_employment_rate, color = MSA)) +
  geom_line() +
  xlim(2019, 2022) +
  theme(legend.position = "none") +
  labs(x = "Year",
       y = "Employment Rate",
      title = "Employment Rate Trend since COVID-19 Pandemic")

# year over year change in employment rate, split up by year
ggplot(MSA_employment, aes(x = avg_emp_rate, y = reorder(MSA, avg_emp_rate))) +
  geom_col(position = "identity", aes(fill = emp_pos), size = 0.25) +
  scale_fill_manual(values = c("red", "darkolivegreen3"), guide = FALSE) +
  facet_grid(~year) +
  labs(x = "Change in employment rate",
       y = " ")

```

2021 Median Rent from <https://www.huduser.gov/portal/datasets/50per.html#2021>

```{r median rent}
# data wrangling 2021 median rent
median_rent2021_data <- read_csv(here("data/raw/median_rent2021.csv")) %>% 
  separate(col = cbsasub21,
           into = c("metro", "cbsa"),
           sep = 5) %>% 
  separate(col = cbsa,
           into = c("unknown", "cbsa"),
           sep = -5) %>% 
  mutate(cbsa = ifelse(cbsa == 'M6400', "38860", cbsa),
        cbsa = as.numeric(cbsa) ,
        median_rent2021 = as.numeric(rent50_1)) %>% # select median rent for 1 bedroom apartments
  select(cbsa, median_rent2021) 

# 2022 median rent
median_rent2022_data <- read_csv(here("data", "raw", "median_rent2022.csv")) %>% 
  separate(col = cbsasub22,
           into = c("metro", "cbsa"),
           sep = 5) %>% 
  separate(col = cbsa,
           into = c("unknown", "cbsa"),
           sep = -5) %>% 
  mutate(cbsa = ifelse(unknown == "38860M" & areaname22 == "Portland, ME HUD Metro FMR Area", "38860", cbsa),
         cbsa = as.numeric(cbsa),
         median_rent2022 = as.numeric(rent_50_1)) %>% 
  select(cbsa, areaname22, median_rent2022) %>% 
  left_join(median_rent2021_data)

client_msas <-client_MSAs %>% 
  left_join(median_rent2022_data)

unfiltered_msas <- unfiltered_msas %>% 
  left_join(median_rent2022_data)
  
```

### House Prices

Median housing prices by MSA were obtained from the Association of Realtors for 2021. Copyright ©[2021] NATIONAL ASSOCIATION OF REALTORS®. All rights reserved. Reprinted with permission.

Source: [The National Association of Realtors](https://www.nar.realtor/research-and-statistics/housing-statistics/metropolitan-median-area-prices-and-affordability)

```{r house price data}
metro_house_cost_data <- read_csv(here("data/raw/metro-home-prices-2022.csv")) %>% 
  row_to_names(3) %>% # make sure column headers are correct
  clean_names() %>% 
  separate(col = "metropolitan_area",
           into = c("cbsa", 'metro_area'),
           sep = 5) %>% 
  mutate(cbsa = as.numeric(cbsa),
        x2021 = as.numeric(x2021), 
        house_price_2021 = x2021 * 1000) %>% # data is proivded as thousands, so multiply to get correct ratio
  mutate(house_price_2021 = ifelse(cbsa == "12260", "284900", house_price_2021), 
        house_price_2021 = ifelse(cbsa == "32580", "259000", house_price_2021)) %>% 
  select(cbsa, house_price_2021)

potential_MSAs <- client_msas %>% 
  left_join(metro_house_cost_data)

unfiltered_msas <- unfiltered_msas %>% 
  left_join(metro_house_cost_data)

```



### MSA data table

```{r data table}
# clean up names
potential_MSAs <- potential_MSAs %>% 
  select(-area, -areaname22)

unfiltered_msas <- unfiltered_msas %>% select(-area, -areaname22)

# table seperated by MSAs
datatable(potential_MSAs, extensions = 'Buttons', options = list(
    dom = 'Bfrtip',
    buttons = c('copy', 'csv', 'excel', 'pdf', 'print')
  )
)

#write_csv(potential_MSAs, here("data/intermediate/demographic_data.csv"))

```
