==> src/03_input_weights.Rmd <==
---
title: "Inputing Weights"
author: "Grace Bianchi"
date: "2023-05-04"
output: 
  html_document:
    code_folding: hide
runtime: shiny
---

```{r setup, include=FALSE, echo=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
library(tidyverse)
library(here)
library(janitor)
library(xlsx)
library(readxl)
```


```{r echo=FALSE}
# list of MSAs without filtering population size
unfiltered_msas <- read_csv(here("data/intermediate/unfilter_msas.csv"))

real_estate_metrics <- read_csv(here("data/intermediate/n_real_estate_metrics.csv")) %>% 
  select(cbsa, anchor_city, msa, starts_with("n_")) %>% 
  mutate(cbsa = factor(cbsa))
  
landlord_metrics <- read_csv(here("data/intermediate/landlord_data.csv")) %>% 
  select(cbsa, starts_with("n_"))%>% 
  mutate(cbsa = factor(cbsa))

climate_risk_metrics <- read_csv(here("data/intermediate/climate_risk_metrics.csv")) %>% 
  mutate(cbsa = factor(cbsa)) %>% 
  select(!risk_score)

reopt_metrics <- read_csv(here("data/intermediate/reopt_metrics")) %>% 
  select(cbsa, starts_with("n_")) %>% 
  mutate(cbsa = factor(cbsa))

```

## Input Weights

### Metric Inputs for Real Estate and Landlord Criteria

```{r real estate and landlord metric weights, echo=FALSE}

real_estate_weights <- c(0.35, # pop change weight
                         0.20, # employment change weight
                         0.15, # occupancy weight
                         0.10, # rent change weight
                         0.10, # rent to income weight
                         0.10) # income to homeprice weight

#input landlord weights for each indicator in the following order: eviction notice days, security deposit limit
# Weights must sum to 1
landlord_weights <- c(0.8, # eviction notice weight
                      0.2) # security deposit limit weight

```

```{r criteria weights, echo=FALSE}
# weights for each criteria must be in this order
equity_weights <- c(0.18, # real estate weight
                    0.0, # landlord weight
                    0.10, # financial weight
                    0.19, # co2 emissions avoided weight
                    0.17, # health impacts weight
                    0.14, # electricity weight
                    0.22) # climate risk weight



```

# Weight Metrics

```{r echo=FALSE}
# weight real estate metrics
wt_real_estate_metrics <- real_estate_metrics %>% 
  mutate(wt_pop_change = n_pop_growth* real_estate_weights[1],
         wt_employ_change = n_avg_emp_change* real_estate_weights[2],
         wt_occ = n_occ_rate2021* real_estate_weights[3],
         wt_rent_change = n_rent_change* real_estate_weights[4],
         wt_rent_income = n_rent_to_income* real_estate_weights[5],
         wt_income_homeprice = n_income_to_homeprice* real_estate_weights[6]) %>% 
  select(!starts_with("n_")) %>% 
  mutate(real_estate_score = apply(.[,4:9], 1, sum)) %>% 
  mutate_if(is.numeric,round, digits = 4) 


# wt_landlord metrics
wt_landlord_metrics <- landlord_metrics %>%
      # multiply each criteria by respective weight
      mutate(wt_eviction = n_eviction_notice*landlord_weights[1],
             wt_deposit = n_security_deposit*landlord_weights[2]) %>% 
      # for loop to sum weighted columns for each row
      mutate(landlord_score = apply(.[, 4:5], 1, sum)) %>% 
   mutate_if(is.numeric,round, digits = 4)  %>% 
  select(!starts_with("n_")) 

```

# Weight Criteria

```{r weighted score function, echo=FALSE}
  calculate_wt_criteria <- function(data, weights){
    wt_criteria <- data %>%
      # multiply each criteria by respective weight
      mutate(wt_real_estate = real_estate_score * weights[1],
             wt_landlord_policy = landlord_score * weights[2], 
             wt_solar_irr = financial_score * weights[3],
            wt_co2 = co2_score * weights[4], 
             wt_health_impacts = health_impact_score * weights[5],
             wt_electricity = electricity_score * weights[6],
             wt_climate_risk = climate_risk_score * weights[7]) %>% 
      # select weighted columns and rename them
      select(!c(ends_with("_score"))) %>%
      # create column for investment score
      mutate(investment_score = 0) %>%
      # for loop to sum weighted columns for each row
      mutate(investment_score = apply(.[, 6:12], 1, sum)) %>% 
      rename("Real Estate"="wt_real_estate",
             "Landlord Policy"="wt_landlord_policy",
             "Electricity Generation"="wt_electricity", 
             "Avoided CO2 Emissions"="wt_co2",
             "Climate Risk Avoided"="wt_climate_risk",
             "Avoided Health Costs"="wt_health_impacts", 
             "Solar Financials"="wt_solar_irr")
  }
  
```


```{r echo=FALSE}
unwt_criteria <- wt_real_estate_metrics %>% 
  inner_join(wt_landlord_metrics) %>% 
  select(!starts_with("wt_")) %>% 
  inner_join(reopt_metrics) %>% 
  inner_join(climate_risk_metrics, by = join_by(cbsa==cbsa)) %>% 
  select(!city_msa) %>% 
    rename(financial_score = n_irr,
         co2_score = n_total_co2,
         health_impact_score = n_health_emissions_cost,
         electricity_score = n_electricity_generation,
         climate_risk_score = n_climate_risk)  %>% 
  mutate_if(is.numeric,round, digits = 4) %>% 
  select(cbsa, msa, anchor_city, state, stateabbrv, ends_with("score"))

equity_test <- calculate_wt_criteria(unwt_criteria, equity_weights)


```

# Visualize results
```{r echo=FALSE}
# data viz
client_score_tidy <- client_test  %>% 
  pivot_longer(cols = c(6:12), # selecting criteria cols
               names_to = "criteria",
               values_to = "wt_criteria_score") %>% 
  unite("msa_state", anchor_city, stateabbrv, sep = ", ") %>% 
 filter(cbsa != "38860")   # removing portland (negative number)
      

ggplot(data = client_score_tidy, aes(y = reorder(msa_state,  wt_criteria_score, FUN=sum), x = wt_criteria_score)) +
    geom_col(aes(fill = factor(criteria, levels = c("Climate Risk Avoided", "Avoided CO2 Emissions", "Electricity Generation", "Solar Financials", "Real Estate", "Landlord Policy", "Avoided Health Costs")))) +
  theme_bw() +
  labs(x = "Investment Favorabliity Score",
       y = " ",
       title = "") +
  guides(fill=guide_legend(title="Criteria (weighted)")) +
   scale_fill_manual(values = c("palevioletred1","orange",  "gold", "#44AA99", "dodgerblue4", "#BC80BD", "#6495ED"),
      breaks = c("Climate Risk Avoided", "Avoided CO2 Emissions", "Electricity Generation", "Solar Financials", "Real Estate", "Landlord Policy", "Avoided Health Costs")) +
        scale_x_continuous(limits = c(0, 0.8), expand = c(0,0), breaks=seq(0, 0.8, 0.1)) 
```
